refactor: remove policyreport package
